\name{MPOmetadata}
\alias{MPOmetadata}
\title{Annotation of MPO Meta Data}
\description{
    This data set gives information of MPO meta data.
}
\details{
    Mappings were based on data provided by: Mouse Phenotype Ontology
    With a date stamp from the source of: 20230302
}

\examples{
    # Convert the object to a data.frame
    library(AnnotationDbi)
    xx <- toTable(MPOmetadata)
}
\references{
\url{http://www.informatics.jax.org/downloads/reports/index.html#pheno}
}
\keyword{datasets}

