\name{MPOMPMGI}
\alias{MPOMPMGI}
\title{Annotation of  MPO Ids to Gene ENTREZID Ids} 
\description{
    This data set describes associations between MPO 
    ids and gene ENTREZID ids, based on the MGI database. 
}
\details{
    Each MPO id is mapped to a vector of gene ENTREZID ids.
    Mappings were based on data provided by: MGI database.
}

\examples{
    # Convert the object to a list
    xx <- as.list(MPOMPMGI)
}

\references{
\url{http://www.informatics.jax.org/downloads/reports/index.html#pheno} 
}
\keyword{datasets}
