\name{MPOOBSOLETE}
\alias{MPOOBSOLETE}
\title{Annotation of MPO Ids by terms defined by Mouse Phenotype Ontology
    Consortium and their status are obsolete}
\description{
This is an R object mapping MPO identifiers to the specific
terms in defined by Mouse Phenotype Ontology Consortium and their definition
are obsolete
}
\details{
    All the obsolete MPO terms that are collected in this index will
    no longer exist in other mapping objects.

    Mappings were based on data provided by: Mouse Phenotype Ontology
    With a date stamp from the source of: 20230302


}

\value{
    \code{MPO_dbconn}: a DBIConnection object representing an
    open connection to the package annotation DB.

    \code{MPO_dbfile}: a character string with the path to the
    package annotation DB.

    \code{MPO_dbschema}: none (invisible \code{NULL}).

    \code{MPO_dbInfo}: none (invisible \code{NULL}).
}

\references{
\url{http://www.informatics.jax.org/downloads/reports/index.html#pheno}
}
\keyword{datasets}

